+++ isqlite/cli.py
@@ -181,21 +181,16 @@ def main_delete(db_path, table, pk, *, where, no_confirm):
     """
     with Database(db_path) as db:
         if pk is not None:
-            try:
-                row = db.get_by_pk(table, pk)
-            except Exception:
-                # This can happen, e.g., if a timestamp column is invalidly formatted and
-                # the Python wrapper around sqlite3 chokes trying to convert it to a
-                # datetime.
-                print(
-                    "Unable to fetch row from database, possibly due to validation error."
-                )
-            else:
-                if row is None:
-                    print(f"Row {pk} not found in table {table!r}.")
-                    sys.exit(1)
+            row = db.get_by_pk(table, pk, get_related=True)
+            for key, value in row.items():
+                if isinstance(value, collections.OrderedDict):
+                    row[key] = get_column_as_string(value)
 
-                prettyprint_row(row)
+            if row is None:
+                print(f"Row {pk} not found in table {table!r}.")
+                sys.exit(1)
+
+            prettyprint_row(row)
 
             if not no_confirm:
                 print()