+++ CHANGELOG.md
@@ -17,7 +17,8 @@ Numbers in parentheses after entries refer to issues in the [GitHub issue tracke
 
 
 ## Unreleased
-Nothing yet.
+### Changed
+- `isqlite update` now treats `NULL` as a special value, in line with `isqlite iupdate`.
 
 
 ## [1.4.0] - 2021-11-17
@@ -28,7 +29,6 @@ Nothing yet.
 
 ### Changed
 - `isqlite sql` no longer prints anything if the SQL query returned no rows. (#66)
-- `isqlite update` now treats `NULL` as a special value, in line with `isqlite iupdate`.
 
 ### Fixed
 - Previously, under some circumstances `Database.diff` would identify a column as renamed (e.g., `a --> b`) even though the old column (`a`) was still in the table. This has been fixed. (#67)